move controllers to presentation layer
