+++ lib/features/weather_page/presentation/widgets/subviews/current_weather.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/features/weather_page/application/current_weather_controller.dart';
 import 'package:flutter_app/features/weather_page/domain/weather/weather_data.dart';
-import 'package:flutter_app/features/weather_page/presentation/subviews/city_search_box.dart';
-import 'package:flutter_app/features/weather_page/presentation/subviews/weather_icon_image.dart';
+import 'package:flutter_app/features/weather_page/presentation/controllers/current_weather_controller.dart';
+import 'package:flutter_app/features/weather_page/presentation/widgets/subviews/city_search_box.dart';
+import 'package:flutter_app/features/weather_page/presentation/widgets/subviews/weather_icon_image.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 class CurrentWeather extends ConsumerWidget {
